Revert "Move tabs files to be more idiomatic with the rest of the app"
This reverts commit 47f0a25065344e2ac7ccd04725fdf27721c37c5c.
