+++ src/pages/task/TaskTabs.tsx
@@ -11,8 +11,8 @@ import { usePrevious } from "hooks";
 import { useTabShortcut } from "hooks/useTabShortcut";
 import { TaskTab } from "types/task";
 import { queryString } from "utils";
-import { BuildBaron } from "./taskTabs/BuildBaronAndAnnotations";
-import useBuildBaronVariables from "./taskTabs/BuildBaronAndAnnotations/useBuildBaronVariables";
+import { BuildBaron } from "./taskTabs/BuildBaron";
+import { useBuildBaronVariables } from "./taskTabs/buildBaronAndAnnotations";
 import { ExecutionTasksTable } from "./taskTabs/ExecutionTasksTable";
 import FilesTableTab from "./taskTabs/FileTableTab";
 import { Logs } from "./taskTabs/Logs";
