+++ src/pages/task/taskTabs/BuildBaron.ts
@@ -1,0 +1,5 @@
+import { loadable } from "components/SpruceLoader";
+
+export const BuildBaron = loadable(
+  () => import("./buildBaronAndAnnotations/BuildBaron")
+);
