+++ src/pages/task/taskTabs/ExecutionTasksTable.stories.tsx
@@ -1,6 +1,6 @@
 import { CustomStoryObj, CustomMeta } from "test_utils/types";
 
-import { ExecutionTasksTable } from ".";
+import { ExecutionTasksTable } from "./ExecutionTasksTable";
 
 export default {
   title: "Pages/Task/Table/Execution Tasks Table",
