+++ src/pages/task/taskTabs/Logs.tsx
@@ -16,7 +16,13 @@ import { size } from "constants/tokens";
 import { TaskLogLinks } from "gql/generated/types";
 import { useUpdateURLQueryParams } from "hooks";
 import { LogTypes, QueryParams } from "types/task";
-import { EventLog, AgentLog, SystemLog, TaskLog, AllLog } from "./LogTypes";
+import {
+  EventLog,
+  AgentLog,
+  SystemLog,
+  TaskLog,
+  AllLog,
+} from "./logs/LogTypes";
 
 const DEFAULT_LOG_TYPE = LogTypes.Task;
