+++ src/pages/task/taskTabs/buildBaronAndAnnotations/index.ts
@@ -1,3 +1,3 @@
-import { loadable } from "components/SpruceLoader";
+import useBuildBaronVariables from "./useBuildBaronVariables";
 
-export const BuildBaron = loadable(() => import("./BuildBaron"));
+export { useBuildBaronVariables };
